# opt/vpp-restapi/docs/swagger.yaml
...
info:
  contact: {}
paths:
  /vpp/bonds:
    get:
      description: List all bond interfaces.
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List Bond Interfaces
      tags:
      - bonds
    post:
      consumes:
      - application/json
      description: Create a new bond interface with members.
      parameters:
      - description: 'Bond Config {mode: string, interfaces: []int}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create Bond Interface
      tags:
      - bonds
  /vpp/bonds/{sw_if_index}:
    delete:
      description: Delete a bond interface by sw_if_index.
      parameters:
      - description: Bond Interface Index
        in: path
        name: sw_if_index
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete Bond Interface
      tags:
      - bonds
  /vpp/bonds/{sw_if_index}/member:
    post:
      consumes:
      - application/json
      description: Add a member to a bond interface.
      parameters:
      - description: Bond Interface Index
        in: path
        name: sw_if_index
        required: true
        type: integer
      - description: 'Member {member_index: int}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add Bond Member
      tags:
      - bonds
...
swagger: "2.0"